Fixed 64-bit CFG recovery
Statement::get_offset() signature had an int return type, while the
offset type was obviously uint64_t.
Fixes #1

+++ src/disassembler/statement.cpp
@@ -5,7 +5,7 @@ Statement::Statement() : offset(0x0), args_at(0)
 {
 }
 
-int Statement::get_offset() const
+uint64_t Statement::get_offset() const
 {
   return offset;
 }
